Add OrderArchive to routes
